Support automatic beacon broadcasting
- Refactor broadcast/scanning to separate lib
- Fix no exposure exception
- Clear User.last_check when resetting exposures
