+++ lib/beacon_history.dart
@@ -1,93 +1,11 @@
 import 'dart:async';
 
-import 'package:beacon_broadcast/beacon_broadcast.dart';
+import 'package:covidtrace/helper/beacon.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_beacon/flutter_beacon.dart';
-import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:intl/intl.dart';
 
 import 'storage/beacon.dart';
-
-// COVID Trace Beacon UUID
-const String UUID = '9F9D2C7D-5022-4052-A36B-B225DBC5E6D2';
-List<Region> regions = [
-  Region(identifier: 'com.covidtrace.app', proximityUUID: UUID)
-];
-
-final BeaconBroadcast beaconBroadcast = BeaconBroadcast();
-StreamSubscription<RangingResult> _streamRanging;
-
-Future<void> setupBeaconScanning() async {
-  try {
-    var success = await flutterBeacon.initializeScanning;
-    print('initScan $success');
-  } catch (err) {
-    print(err);
-  }
-
-  // Start monitoring for entry/exit of a region
-  flutterBeacon.monitoring(regions).listen((MonitoringResult result) async {
-    print('monitoring... $result');
-    print(result.monitoringEventType);
-    print(result.monitoringState);
-    print(result.region);
-
-    switch (result.monitoringState) {
-      case MonitoringState.inside:
-        var advertising = await beaconBroadcast.isAdvertising();
-        if (!advertising) {
-          showBeaconNotification();
-        }
-        startRegionRange();
-        break;
-      case MonitoringState.outside:
-        stopRegionRange();
-        break;
-      case MonitoringState.unknown:
-        break;
-    }
-  });
-}
-
-void startRegionRange() {
-  print('starting region range');
-  _streamRanging =
-      flutterBeacon.ranging(regions).listen((RangingResult result) async {
-    // TODO(wes): Rotate clientId if there is a collision detected
-    Future.wait(
-        result.beacons.map((b) => BeaconTransmission.seen(b.major, b.minor)));
-    await cleanupTransmissions();
-  });
-}
-
-Future<void> cleanupTransmissions() async {
-  var count = await BeaconTransmission.endUnseen();
-  if (count > 0) {
-    await BeaconTransmission.convertCompleted();
-  }
-}
-
-Future<void> stopRegionRange() async {
-  print('stopping region range');
-  _streamRanging?.cancel();
-  await Future.delayed(BeaconTransmission.UNSEEN_TIMEOUT);
-  await cleanupTransmissions();
-}
-
-void showBeaconNotification() async {
-  var notificationPlugin = FlutterLocalNotificationsPlugin();
-  var androidSpec = AndroidNotificationDetails(
-      '1', 'COVID Trace', 'Beacon notification',
-      importance: Importance.Max, priority: Priority.High, ticker: 'ticker');
-  var iosSpecs = IOSNotificationDetails();
-  await notificationPlugin.show(
-      1,
-      'COVID Trace Monitoring Alert',
-      'Keep the app open to get accurate monitoring while indoors.',
-      NotificationDetails(androidSpec, iosSpecs),
-      payload: 'Default_Sound');
-}
 
 class BeaconHistory extends StatefulWidget {
   @override
@@ -97,7 +15,6 @@ class BeaconHistory extends StatefulWidget {
 class BeaconState extends State {
   BeaconUuid _beacon;
   bool _broadcasting = false;
-  bool _advertising = false;
   List<BeaconTransmission> _transmissions = [];
   List<BeaconModel> _beacons = [];
   Timer refreshTimer;
@@ -109,83 +26,31 @@ class BeaconState extends State {
   void initState() {
     super.initState();
 
-    initBroadcast();
     refreshBeacons();
-
     refreshTimer = Timer.periodic(Duration(seconds: 1), (timer) {
       refreshBeacons();
     });
+
+    setState(() => _broadcasting = isBroadcasting());
   }
 
   @override
   void dispose() {
     super.dispose();
     refreshTimer.cancel();
-    stopAdvertising();
-  }
-
-  void initBroadcast() {
-    beaconBroadcast.getAdvertisingStateChange().listen((isAdvertising) {
-      if (mounted) {
-        setState(() => _advertising = isAdvertising);
-      }
-    });
-
-    beaconBroadcast
-        .setUUID(UUID)
-        .setIdentifier('com.covidtrace.app')
-        .setLayout('m:2-3=0215,i:4-19,i:20-21,i:22-23,p:24-24') // iBeacon
-        .setManufacturerId(0x004C); // Apple
-  }
-
-  void onBroadcastChange(bool value) async {
-    setState(() => _broadcasting = value);
-    if (value) {
-      await startAdvertising();
-    } else {
-      stopAdvertising();
-    }
-  }
-
-  Future<void> startAdvertising() async {
-    var beacon = await BeaconUuid.get();
-    setState(() => _beacon = beacon);
-
-    sequenceTimer?.cancel();
-    sequenceTimer = Timer.periodic(Duration(seconds: 4), (timer) async {
-      if (_advertising) {
-        beaconBroadcast.stop();
-        await Future.delayed(Duration(seconds: 2));
-      }
-
-      if (!timer.isActive || !_broadcasting) {
-        return;
-      }
-
-      await _beacon.next();
-      beaconBroadcast.setMajorId(_beacon.major);
-      beaconBroadcast.setMinorId(_beacon.minor);
-
-      if (!_advertising) {
-        beaconBroadcast.start();
-      }
-    });
-  }
-
-  void stopAdvertising() {
-    sequenceTimer?.cancel();
-    beaconBroadcast.stop();
   }
 
   Future<void> refreshBeacons() async {
     var transmissions = await BeaconTransmission.findAll(orderBy: 'start DESC');
     var beacons = await BeaconModel.findAll(orderBy: 'start DESC');
+    var beacon = getBeaconUuid();
 
     if (!mounted) {
       return;
     }
 
     setState(() {
+      _beacon = beacon;
       _beacons = beacons;
       _transmissions = transmissions;
     });
@@ -202,6 +67,15 @@ class BeaconState extends State {
     await BeaconTransmission.destroy(
         where: 'clientId = ? AND last_seen = ? AND end IS NOT NULL',
         whereArgs: [clientId, lastSeen.toIso8601String()]);
+  }
+
+  void onBroadcastChange(bool value) async {
+    if (value) {
+      await startAdvertising();
+    } else {
+      stopAdvertising();
+    }
+    setState(() => _broadcasting = isBroadcasting());
   }
 
   @override
@@ -284,7 +158,7 @@ class BeaconState extends State {
                 iconColor: Colors.white,
                 child: ListTile(
                     title: Text('Broadcasting'),
-                    subtitle: _beacon != null && _broadcasting
+                    subtitle: _broadcasting && _beacon != null
                         ? Text(
                             '${_beacon.clientId}.${_beacon.offset}.${_beacon.major}')
                         : Text('is turned off'),
