+++ lib/onboarding.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'package:app_settings/app_settings.dart';
+import 'package:covidtrace/helper/beacon.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -9,7 +10,6 @@ import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import 'beacon_history.dart';
 import 'helper/location.dart';
 import 'storage/location.dart';
 import 'storage/user.dart';
@@ -100,6 +100,7 @@ class OnboardingState extends State {
         bg.BackgroundGeolocation.start();
         // TODO(wes): This will immediately prompt for bluetooth. Should we move this somewhere else?
         setupBeaconScanning();
+        setupBeaconBroadcast();
       } else {
         setState(() => _linkToSettings = true);
       }
